backend/requirements.txt
# ============================================================
# Web / API Core
# ============================================================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
starlette>=0.27.0

# ============================================================
# LLM / AI
# ============================================================
openai>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.0
torch>=2.8.0
torchvision>=0.23.0
torchaudio>=2.8.0

# ============================================================
# Configuration / Settings
# ============================================================
python-dotenv>=1.2.1
pydantic>=2.12.5
pydantic-settings>=2.1.0
PyYAML>=6.0.1

# ============================================================
# Storage / Database (SQLite)
# ============================================================
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ============================================================
# Caching / Lightweight State
# ============================================================
redis>=5.0.0
hiredis>=2.0.0

# ============================================================
# Security / Auth
# ============================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
cryptography>=44.0.0

# ============================================================
# Networking / HTTP
# ============================================================
requests>=2.31.0
httpx>=0.27.0

# ============================================================
# Utilities
# ============================================================
psutil>=7.2.1
gputil>=1.4.0
onnxruntime>=1.19.0
orjson>=3.9.0
jsonschema>=4.21.0
ddgs>=9.10.0

# ============================================================
# Testing
# ============================================================
pytest>=9.0.2
pytest-asyncio>=1.3.0
respx>=0.22.0
